Servico.java: mudando o argumento int valor do construtor para float valor; Main.java: adequando valor de servico para float; Funcoes.java: Arrumando o bug (equals ao inves de == para comparacao de strings

+++ Main.java
@@ -118,14 +118,14 @@ public class Main {
 
 			else if(comando.equals("4")) {
 				String nome;
-				int valor;
+				float valor;
 
 				System.out.println("Insira os dados do servico");
 				System.out.println("Nome:");
 				nome = input.nextLine();
 				System.out.println("Valor:");
-				valor = Integer.parseInt(input.nextLine());
-
+				valor = Float.parseFloat(input.nextLine());
+				
 				servicos.add(new Servico(nome, valor));
 			}
